Move loyalty related components under 'checkout' package
