+++ src/main/java/com/github/jbence1994/webshop/Application.java
@@ -1,11 +1,11 @@
 package com.github.jbence1994.webshop;
 
 import com.github.jbence1994.webshop.checkout.FreeShippingConfig;
+import com.github.jbence1994.webshop.checkout.LoyaltyConfig;
 import com.github.jbence1994.webshop.common.ClientAppConfig;
 import com.github.jbence1994.webshop.common.WebshopEmailAddressConfig;
 import com.github.jbence1994.webshop.common.WebshopNameConfig;
 import com.github.jbence1994.webshop.image.ImageUploadsConfig;
-import com.github.jbence1994.webshop.loyalty.LoyaltyConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
