+++ src/main/java/com/github/jbence1994/webshop/checkout/LoyaltyPointsCalculator.java
@@ -1,4 +1,4 @@
-package com.github.jbence1994.webshop.loyalty;
+package com.github.jbence1994.webshop.checkout;
 
 import java.math.BigDecimal;
