+++ src/test/java/com/github/jbence1994/webshop/checkout/LoyaltyTestConstants.java
@@ -1,4 +1,4 @@
-package com.github.jbence1994.webshop.loyalty;
+package com.github.jbence1994.webshop.checkout;
 
 public interface LoyaltyTestConstants {
     int POINTS_RATE = 20;
